feat: add vector image handling for office file conversions
